moved many reusable modular functions out of hdf_writer to write_utils Rewrote build_ind_val_dsets and build_reduced_spec_dsets to NOT use VirtualDataset
