+++ pycroscopy/core/io/hdf_utils.py
@@ -12,8 +12,7 @@ import collections
 from warnings import warn
 from collections import Iterable
 import numpy as np
-from .virtual_data import VirtualDataset
-from .write_utils import make_indices_matrix, INDICES_DTYPE, VALUES_DTYPE
+from .write_utils import make_indices_matrix, INDICES_DTYPE
 
 __all__ = ['get_attr', 'get_h5_obj_refs', 'get_indices_for_region_ref', 'get_dimensionality', 'get_sort_order',
            'get_auxillary_datasets', 'get_attributes', 'get_group_refs', 'check_if_main', 'check_and_link_ancillary',
@@ -22,7 +21,7 @@ __all__ = ['get_attr', 'get_h5_obj_refs', 'get_indices_for_region_ref', 'get_dim
            'link_h5_obj_as_alias',
            'find_results_groups', 'get_formatted_labels', 'reshape_from_n_dims', 'find_dataset', 'print_tree',
            'copy_main_attributes', 'create_empty_dataset', 'calc_chunks', 'create_spec_inds_from_vals',
-           'build_reduced_spec_dsets', 'check_for_old', 'get_source_dataset', 'get_unit_values', 'get_data_descriptor',
+           'check_for_old', 'get_source_dataset', 'get_unit_values', 'get_data_descriptor',
            'link_as_main', 'copy_reg_ref_reduced_dim', 'simple_region_ref_copy']
 
 if sys.version_info.major == 3:
@@ -191,7 +190,6 @@ def get_attributes(h5_object, attr_names=None):
     -------
     Dictionary containing (name,value) pairs of attributes
     """
-    from .pycro_data import PycroDataset
 
     assert isinstance(h5_object, (h5py.File, h5py.Group, h5py.Dataset))
     if attr_names is not None:
@@ -1485,76 +1483,6 @@ def simple_region_ref_copy(h5_source, h5_target, key):
     target_ref = create_region_reference(h5_target, ref_inds)
     h5_target.attrs[key] = target_ref
     return ref_inds
-
-
-def build_reduced_spec_dsets(h5_spec_inds, h5_spec_vals, keep_dim, step_starts, basename='Spectroscopic'):
-    """
-    Creates new Spectroscopic Indices and Values datasets from the input datasets
-    and keeps the dimensions specified in keep_dim
-
-    Parameters
-    ----------
-    h5_spec_inds : HDF5 Dataset
-            Spectroscopic indices dataset
-    h5_spec_vals : HDF5 Dataset
-            Spectroscopic values dataset
-    keep_dim : Numpy Array, Boolean
-            Array designating which rows of the input spectroscopic datasets to keep
-    step_starts : Numpy Array, Unsigned Integers
-            Array specifying the start of each step in the reduced datasets
-    basename : String
-            String to which '_Indices' and '_Values' will be appended to get the names
-            of the new datasets
-
-    Returns
-    -------
-    ds_inds : VirtualDataset
-            Reduced Spectroscopic indices dataset
-    ds_vals : VirtualDataset
-            Reduces Spectroscopic values dataset
-    """
-    for param in [h5_spec_inds, h5_spec_vals]:
-        assert isinstance(param, h5py.Dataset)
-    assert isinstance(keep_dim, (bool, np.ndarray))
-    assert isinstance(step_starts, (list, np.ndarray))
-    if basename is not None:
-        assert isinstance(basename, (str, unicode))
-
-    if h5_spec_inds.shape[0] > 1:
-        '''
-        Extract all rows that we want to keep from input indices and values
-        '''
-        ind_mat = h5_spec_inds[keep_dim, :][:, step_starts]
-        val_mat = h5_spec_vals[keep_dim, :][:, step_starts]
-        '''
-        Create new VirtualDatasets to hold the data
-        Name them based on basename
-        '''
-        ds_inds = VirtualDataset(basename + '_Indices', ind_mat, dtype=h5_spec_inds.dtype)
-        ds_vals = VirtualDataset(basename + '_Values', val_mat, dtype=h5_spec_vals.dtype)
-        # Extracting the labels from the original spectroscopic data sets
-        sho_inds_labs = h5_spec_inds.attrs['labels'][keep_dim]
-        # Creating the dimension slices for the new spectroscopic data sets
-        inds_slices = dict()
-        for row_ind, row_name in enumerate(sho_inds_labs):
-            inds_slices[row_name] = (slice(row_ind, row_ind + 1), slice(None))
-
-        # Adding the labels and units to the new spectroscopic data sets
-        ds_inds.attrs['labels'] = inds_slices
-        ds_inds.attrs['units'] = h5_spec_inds.attrs['units'][keep_dim]
-        ds_vals.attrs['labels'] = inds_slices
-        ds_vals.attrs['units'] = h5_spec_vals.attrs['units'][keep_dim]
-
-    else:  # Single spectroscopic dimension:
-        ds_inds = VirtualDataset('Spectroscopic_Indices', np.array([[0]], dtype=INDICES_DTYPE))
-        ds_vals = VirtualDataset('Spectroscopic_Values', np.array([[0]], dtype=VALUES_DTYPE))
-
-        ds_inds.attrs['labels'] = {'Single_Step': (slice(0, None), slice(None))}
-        ds_vals.attrs['labels'] = {'Single_Step': (slice(0, None), slice(None))}
-        ds_inds.attrs['units'] = ['']
-        ds_vals.attrs['units'] = ['']
-
-    return ds_inds, ds_vals
 
 
 def calc_chunks(dimensions, dtype_byte_size, unit_chunks=None, max_chunk_mem=10240):
