+++ pycroscopy/io/translators/beps_data_generator.py
@@ -8,7 +8,8 @@ from sklearn.utils import gen_batches
 from skimage.measure import block_reduce
 # Pycroscopy imports
 from ...core.io.hdf_writer import HDFwriter
-from ...core.io.hdf_utils import calc_chunks, get_h5_obj_refs, link_as_main, get_attr, build_reduced_spec_dsets
+from ...core.io.hdf_utils import calc_chunks, get_h5_obj_refs, link_as_main, get_attr
+from pycroscopy.core.io.write_utils import build_reduced_spec_dsets
 from ...core.io.dtype_utils import stack_real_to_compound
 from ...core.io.translator import Translator, generate_dummy_main_parms
 from ...core.io.write_utils import build_ind_val_dsets
